web-animations/interfaces/AnimationEffectTiming/iterationStart.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>AnimationEffectTiming.iterationStart</title>
<link rel="help" href="https://drafts.csswg.org/web-animations/#dom-animationeffecttiming-iterationstart">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../testcommon.js"></script>
<script src="../../resources/timing-tests.js"></script>
<body>
<div id="log"></div>
<script>
'use strict';

test(t => {
  const anim = createDiv(t).animate(null,
                                    { iterationStart: 0.2,
                                      iterations: 1,
                                      fill: 'both',
                                      duration: 100,
                                      delay: 1 });
  anim.effect.timing.iterationStart = 2.5;
  assert_time_equals_literal(anim.effect.getComputedTiming().progress, 0.5);
  assert_equals(anim.effect.getComputedTiming().currentIteration, 2);
}, 'Allows setting the iterationStart of an animation in progress:'
   + ' backwards-filling');

test(t => {
  const anim = createDiv(t).animate(null,
                                    { iterationStart: 0.2,
                                      iterations: 1,
                                      fill: 'both',
                                      duration: 100,
                                      delay: 0 });
  anim.effect.timing.iterationStart = 2.5;
  assert_time_equals_literal(anim.effect.getComputedTiming().progress, 0.5);
  assert_equals(anim.effect.getComputedTiming().currentIteration, 2);
}, 'Allows setting the iterationStart of an animation in progress:'
   + ' active phase');

test(t => {
  const anim = createDiv(t).animate(null,
                                    { iterationStart: 0.2,
                                      iterations: 1,
                                      fill: 'both',
                                      duration: 100,
                                      delay: 0 });
  anim.finish();
  anim.effect.timing.iterationStart = 2.5;
  assert_time_equals_literal(anim.effect.getComputedTiming().progress, 0.5);
  assert_equals(anim.effect.getComputedTiming().currentIteration, 3);
}, 'Allows setting the iterationStart of an animation in progress:'
   + ' forwards-filling');

for (const invalid of gBadIterationStartValues) {
  test(t => {
    const anim = createDiv(t).animate(null);
    assert_throws({ name: 'TypeError' }, () => {
      anim.effect.timing.iterationStart = invalid;
    }, `setting ${invalid}`);
  }, `Throws when setting invalid iterationStart value: ${invalid}`);
}

</script>
</body>
